Address PR #42 code review feedback
## Critical Fixes

### Physics Bugs
- **Gravity force**: Removed incorrect mass division - all objects now fall at same rate regardless of mass
- **Particle fading**: Fixed negative float to uint8_t cast bug that caused intensity jumps instead of fade
- **Position update**: Removed incorrect 30.0f multiplier - velocities now properly use LEDs/sec units
- **Drag force**: Replaced expensive pow() with linear approximation for better performance

### Memory Management
- **activeCount_ tracking**: Fixed double-decrement bug by tracking count before update
- Prevents pool state corruption when particles are explicitly killed during update

### Code Quality
- **Redundant field**: Removed maxParticles from all generator params (Fire, Water, Lightning)
  - Pool size is determined by template parameter, not runtime field
  - Removed from param structs, SerialConsole registration, and ConfigStorage load/save
- **Overflow protection**: Added uint32_t cast in Water splash intensity calculation
- **Unused include**: Removed <string.h> from ParticlePool.h

## Documentation Updates
- Updated Particle.h velocity units comment (LEDs/sec not LEDs/frame@30FPS)
- Updated force application comments for clarity

## Skipped (Requires Larger Refactoring)
- ParticleFlags enum class conversion - would require updating all bitwise operators and usage sites
- Can be addressed in separate PR if type safety improvement is desired

## Compilation Status
✅ All fixes compile successfully (190,840 bytes, 23% flash usage)
✅ No warnings or errors

+++ blinky-things/generators/Water.cpp
@@ -110,7 +110,7 @@ void Water::spawnSplash(float x, float y, uint8_t parentIntensity) {
         float vx = cos(angle) * speed;
         float vy = sin(angle) * speed - 1.0f;  // Slight upward component
 
-        uint8_t intensity = min(255, parentIntensity * params_.splashIntensity / 255);
+        uint8_t intensity = min(255, (uint32_t)parentIntensity * params_.splashIntensity / 255);
 
         pool_.spawn(x, y, vx, vy, intensity, 30, 0.5f,  // Light, short-lived
                    ParticleFlags::GRAVITY | ParticleFlags::FADE);
